[Move #234] MainCreateRoom 폴더 이름 main으로 변경
